refactor(YouTube - Download Playlist Button): Move to separate patch
